fft function prototype
